// frontend-recovered/mansis-admin-original-ghcr-latest/html/assets/recovered-sources/index-DsuMR42Q.js/src/enums/purchase-type.ts
export enum PurchaseType {
  CASH = 'CASH',
  WALLET = 'WALLET',
  CAMPAIGN = 'CAMPAIGN',
  CARD = 'CARD',
  DIRECT = 'DIRECT',
  MEMBERSHIP = 'MEMBERSHIP',
  PHYSICAL_CARD = 'PHYSICAL_CARD',
  LOYALTY_POINTS = 'LOYALTY_POINTS',
  LOYALTY_POINTS_HYBRID = 'LOYALTY_POINTS_HYBRID',
  MIXED = 'MIXED'
}